refactor: split presentation module into feature modules
- presentation 모듈을 feature:home, feature:login, feature:groups로 분리
