--- core/ui/src/main/res/values/string.xml
@@ -1,43 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <!-- app -->
-    <string name="point_unit">%1d 아쿠</string>
-
-    <!-- image description -->
-    <string name="char_head_unknown_description">Empty Schedule Illustration</string>
-    <string name="char_running_boy_description">Empty State Illustration</string>
-    <string name="char_head_no_hair_description">Default Profile Illustration</string>
-
-    <!-- icon description -->
-    <string name="ic_location_description">location icon</string>
-
-    <!-- group -->
-    <string name="group_empty_message">그룹을 만들어 보세요!</string>
-    <string name="group_empty_button">+ 그룹 만들기</string>
-    <string name="group_card_recent_schedule">최근 약속 %1$s</string>
-    <string name="group_card_badge_member_over_limit">99+</string>
-
-    <!-- schedule -->
-    <string name="schedule_empty_message">아직 약속이 없어요</string>
-    <string name="schedule_status_running">진행 중</string>
-    <string name="schedule_status_waiting">대기 중</string>
-
-    <!-- dialog -->
-    <string name="create_group_dialog_title">그룹 생성</string>
-    <string name="create_group_dialog_placeholder">그룹 이름을 입력하세요</string>
-    <string name="create_group_dialog_supporting_text">최대 15자 입력이 가능합니다</string>
-    <string name="create_group_dialog_button">생성하기</string>
-    <string name="create_group_dialog_error_blank">그룹명을 입력해주세요</string>
-    <string name="create_group_dialog_error_too_short">2자 이상 입력해주세요</string>
-    <string name="create_group_dialog_error_special_chars">특수문자는 입력 불가합니다</string>
-    <string name="date_picker_dialog_button">확인</string>
-
-    <string name="insufficient_points_dialog_title">야쿠 부족</string>
-    <string name="insufficient_points_dialog_message">야쿠가 부족합니다!\n광고 시청 후 야쿠를 충전해주세요.</string>
-    <string name="insufficient_points_dialog_current_points_label">현재 내 보유 아쿠</string>
-    <string name="insufficient_points_dialog_button">충선소 바로가기</string>
-
-    <string name="join_schedule_dialog_participant_format">%1$s 외 %2$d명 참가</string>
-    <string name="join_schedule_dialog_point_notice">약속 입장 시, 10야쿠가 소모됩니다.</string>
-    <string name="join_schedule_dialog_button">입장하기</string>
-</resources>
