+++ feature/home/src/main/kotlin/com/hyunjung/aiku/feature/home/HomeGroupUiState.kt
@@ -1,0 +1,7 @@
+package com.hyunjung.aiku.feature.home
+
+sealed interface HomeGroupUiState {
+    data object Loading : HomeGroupUiState
+    data object Idle : HomeGroupUiState
+    data class Error(val message: String) : HomeGroupUiState
+}
