+++ feature/home/src/main/kotlin/com/hyunjung/aiku/feature/home/HomeScheduleUiState.kt
@@ -1,0 +1,7 @@
+package com.hyunjung.aiku.feature.home
+
+sealed interface HomeScheduleUiState {
+    data object Loading : HomeScheduleUiState
+    data object Idle : HomeScheduleUiState
+    data class Error(val message: String) : HomeScheduleUiState
+}
