--- presentation/src/main/kotlin/com/hyunjung/aiku/presentation/home/model/HomeGroupUiState.kt
@@ -1,7 +1,0 @@
-package com.hyunjung.aiku.presentation.home.model
-
-sealed interface HomeGroupUiState {
-    data object Loading : HomeGroupUiState
-    data object Idle : HomeGroupUiState
-    data class Error(val message: String) : HomeGroupUiState
-}
