--- presentation/src/main/kotlin/com/hyunjung/aiku/presentation/splash/SplashActivity.kt
@@ -1,135 +1,0 @@
-package com.hyunjung.aiku.presentation.splash
-
-import android.annotation.SuppressLint
-import android.content.Intent
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.SystemBarStyle
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentWidth
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
-import com.hyunjung.aiku.core.designsystem.theme.AiKUTheme
-import com.hyunjung.aiku.core.designsystem.theme.AikuColors
-import com.hyunjung.aiku.core.designsystem.theme.AikuTypography
-import com.hyunjung.aiku.presentation.R
-import com.hyunjung.aiku.presentation.main.MainActivity
-import dagger.hilt.android.AndroidEntryPoint
-import kotlinx.coroutines.delay
-import com.hyunjung.aiku.core.ui.R as UiRes
-
-@SuppressLint("CustomSplashScreen")
-@AndroidEntryPoint
-class SplashActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge(
-            statusBarStyle = SystemBarStyle.light(
-                Color.Transparent.toArgb(),
-                Color.Transparent.toArgb()
-            ),
-            navigationBarStyle = SystemBarStyle.light(
-                Color.Transparent.toArgb(),
-                Color.Transparent.toArgb()
-            )
-        )
-
-        setContent {
-            AiKUTheme {
-                SplashScreen(
-                    onSplashFinished = {
-                        startActivity(Intent(this, MainActivity::class.java).apply {
-                            flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-                        })
-                        finish()
-                    }
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun SplashScreen(
-    onSplashFinished: () -> Unit
-) {
-    LaunchedEffect(Unit) {
-        delay(2000)
-        onSplashFinished()
-    }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(AikuColors.Green05),
-        contentAlignment = Alignment.Center
-    ) {
-        Column(horizontalAlignment = Alignment.CenterHorizontally) {
-            Text(
-                text = stringResource(R.string.presentation_app_title),
-                color = AikuColors.CobaltBlue,
-                style = AikuTypography.Subtitle4_G
-            )
-            Text(
-                text = stringResource(R.string.presentation_app_name),
-                color = AikuColors.CobaltBlue,
-                style = AikuTypography.Headline1_G
-            )
-            Spacer(modifier = Modifier.padding(24.dp))
-            SplashCharactersRow()
-        }
-    }
-}
-
-@Composable
-fun SplashCharactersRow() {
-    Row(
-        modifier = Modifier
-            .wrapContentWidth(align = Alignment.CenterHorizontally, unbounded = true),
-        horizontalArrangement = Arrangement.spacedBy(16.dp),
-        verticalAlignment = Alignment.CenterVertically
-    ) {
-        Image(
-            painter = painterResource(id = UiRes.drawable.img_char_1),
-            contentDescription = "Character 1",
-        )
-        Image(
-            painter = painterResource(id = UiRes.drawable.img_char_2),
-            contentDescription = "Character 2",
-
-            )
-        Image(
-            painter = painterResource(id = UiRes.drawable.img_char_3),
-            contentDescription = "Character 3"
-        )
-        Image(
-            painter = painterResource(id = UiRes.drawable.img_char_4),
-            contentDescription = "Character 4",
-        )
-    }
-}
-
-@Preview
-@Composable
-private fun SplashScreenPreview() {
-    SplashScreen(onSplashFinished = {})
-}
