--- presentation/src/main/res/values/strings.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="presentation_app_name">AiKU</string>
-    <string name="presentation_app_title">아이쿠! 또 지각이네?!</string>
-    <string name="feature_home_schedule_title">오늘 내 약속</string>
-    <string name="feature_home_group_title">%1$s\'s Group</string>
-    <string name="feature_home_fab_add_group">그룹 추가</string>
-    <string name="presentation_member_detail_member_section_empty_title">멤버를 초대해 보세요!</string>
-    <string name="presentation_member_detail_member_section_empty_button">+ 멤버 초대하기</string>
-    <string name="presentation_member_detail_member_section_invite">초대하기</string>
-    <string name="presentation_member_detail_member_section_avatar_description">%1$s의 프로필</string>
-    <string name="presentation_member_detail_schedule_section_empty_title">약속을 만들어 주세요!</string>
-    <string name="presentation_member_detail_schedule_section_empty_button">+ 약속 만들기</string>
-    <string name="presentation_member_detail_schedule_section_schedule_status_summary">진행중 %1$d | 대기중 %2$d</string>
-    <string name="presentation_create_schedule_name_default_supporting_text">최대 15자 입력이 가능합니다</string>
-    <string name="presentation_create_schedule_name_error_blank">약속 이름을 입력해주세요</string>
-    <string name="presentation_create_schedule_name_error_too_short">2자 이상 입력해주세요</string>
-    <string name="presentation_create_schedule_name_error_special_chars">특수문자는 입력 불가합니다</string>
-    <string name="presentation_create_schedule_name_title">약속 이름</string>
-    <string name="presentation_create_schedule_name_placeholder">약속 이름을 입력해 주세요</string>
-    <string name="presentation_create_schedule_datetime_title">약속 날짜・시간</string>
-    <string name="presentation_create_schedule_datetime_supporting">현재시간 기준 40분 이후부터 개설 가능합니다</string>
-    <string name="presentation_create_schedule_location_title">약속 장소</string>
-    <string name="presentation_create_schedule_location_placeholder">약속 장소를 입력해 주세요</string>
-    <string name="presentation_create_schedule_button_create">만들기</string>
-    <string name="presentation_create_schedule_date_description">날짜 선택</string>
-    <string name="presentation_create_schedule_time_description">시간 선택</string>
-    <string name="presentation_create_schedule_location_description">장소 검색</string>
-</resources>
