+++ settings.gradle.kts
@@ -38,4 +38,6 @@ include(":core:navigation")
 include(":core:network")
 include(":core:model")
 include(":core:ui")
-include(":presentation")
+include(":feature:home")
+include(":feature:login")
+include(":feature:groups")
